third_party/blink/web_tests/webcodecs/basic_video_encoding.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<img id='frame_image' style='display: none;' src="pattern.png">
<script>

'use strict';

async function generateBitmap(width, height, text) {
  let img = document.getElementById('frame_image');
  let cnv = document.createElement("canvas");
  cnv.height = height;
  cnv.width = width;
  var ctx = cnv.getContext('2d');
  ctx.drawImage(img, 0, 0);
  ctx.font = '20px serif';
  ctx.fillText(text, 0, 0);
  return createImageBitmap(cnv);
}

async function createFrame(width, height, ts) {
  let imageBitmap = await generateBitmap(width, height, ts.toString());
  return new VideoFrame({ timestamp: ts }, imageBitmap);
}

promise_test(async function() {
  const tune = {
    width: 123,
    height: 321,
    biterate: 1e6,
    framerate: 24
   }
  const init_params = {
    codec : "NoOpCodec",
    tuneOptions: tune,
    output : () => { return true; }
  };
  let encoder = new VideoEncoder(init_params);
  try {
    tune.width = null;
    let encoder = new VideoEncoder(init_params);
  } catch (e) {
    assert_equals(e.name, "ConstraintError");
  }
}, "Initialization");

promise_test(async function() {
  let w = 256;
  let h = 256;
  let frames_to_encode = 20;
  let frames_processed = 0;
  let process_video_chunk = function(chunk) {
    assert_greater_than_equal(chunk.timestamp, 1);
    let type = (chunk.timestamp == 1) ? "key" : "delta";
    assert_equals(chunk.type, type);
    var data = new Uint8Array(chunk.data);
    assert_greater_than_equal(data.length, w);
    frames_processed++;
  };

  const init_params = {
    codec : "NoOpCodec",
    tuneOptions : {
      width : w,
      height : h
    },
    output : process_video_chunk
  };
  let encoder = new VideoEncoder(init_params);
  for (let i = 0; i < frames_to_encode; i++) {
    var frame = await createFrame(w, h, i + 1);
    await encoder.encode(frame, { keyFrame : (i == 0)});
  }
  assert_equals(frames_processed, frames_to_encode);
}, "EncodeFrame");

</script>
</body>
</html>
